+++ CHANGELOG.md
@@ -7,13 +7,16 @@
 
 * support single group and platform values (e.g., `group: :test` or `platform: :mri`)
 
+### Bug Fixes
+
+* `release` workflow: Use proper `/v2/` in path to `golangci-lint` in `magefile.go`
+
 
 ## [0.9.0](https://github.com/contriboss/gemfile-go/compare/v0.8.0...v0.9.0) (2026-02-08)
 
 
 ### Features
 
-* ✨ support hash 🚀 syntax in gemfiles and gemspecs ([a120bcc](https://github.com/contriboss/gemfile-go/commit/a120bcc01003c8e087c053dca2a5dfd50888f262))
 * ✨ support hash 🚀 syntax in gemfiles and gemspecs ([d8f48ee](https://github.com/contriboss/gemfile-go/commit/d8f48eef4ae5bfd3da4113346856133b61a0f1e9))
 
 ### Bug Fixes